moved tests that rely on callback to their own folder
